+++ config.json
@@ -117,7 +117,7 @@
         "prefix": "w"
       }
     ],
-    "plan_ids": [1, 2, 3, 4, 5, 6, 7]
+    "plan_ids": [1, 2, 3, 4, 5, 6, 7, 8]
   },
   "network": {
     "web3_provider_url": "https://polygon-mainnet.g.alchemy.com/v2/6kkkcccOr3uJck57CliIomFbHt9eSYu5",
