+++ pages/cloutpatrons/index.tsx
@@ -1,0 +1,8 @@
+import type { NextPage } from 'next';
+import Home from '../Home';
+
+const Index: NextPage = () => {
+  return <Home />;
+};
+
+export default Index;
